fix: 复用单元格滚动后TableSheet的seriesNumber和普通单元格样式异常
